requirements.txt
numpy==2.3.5
pandas==2.3.3
protobuf==6.33.1
scikit_learn==1.7.2
streamlit==1.51.0
google-genai==1.52.0